gwt: Move CookieManager and ExceptionHandler to the ro.pub.cs.vmchecker.client.util package.
